Adding basic bootstrap assets for site layout.
As part of #5, adding new assets to facilitate making a nice layout.
Also adding a static link to the assets. On the development server, we
can make this cleaner by using a site root and a local_settings.py file.
On the production server, this should be handled by whatever production
server is running (i.e. Apache).