Moving balance out to a derivable function.
In order to facilitate computation of balance, added a get_balance
function to the account structure and removed the hard coded balance
line. Since issue #2 requires that accounts operate on transactions,
this should be a good way of ensuring that balances are properly
computed. May be a good idea to add back in a balance column (derivable
data stored in database) for quick lookup--effectively caching balance
data--so that the system does not need to recompute the balance every
time.