Trying a bit more to fix up signals for #8
Signals now use decorators added in Django 1.3. Because I want to mess
around with some cool new stuff, I'm just upgrading the dependancy
requirement to Django 1.4. I changed the model to save with
update_balance instead of zeroing balance, thinking maybe that was the
issue. It wasn't. What. The. Fuck.